MueLu: Initial refactoring to allow the perf reports to work with the MueLu Driver
